refactor: Feed의 cache 관련 repository로 위치 이동 및 조회 repository는 query repo로 이름 변경
