+++ src/main/java/com/LetMeDoWith/LetMeDoWith/application/feed/service/FeedService.java
@@ -2,7 +2,7 @@ package com.LetMeDoWith.LetMeDoWith.application.feed.service;
 
 import com.LetMeDoWith.LetMeDoWith.application.feed.dto.RetrieveFeedbackAvailableDowithTasksResult;
 import com.LetMeDoWith.LetMeDoWith.common.util.SystemTimeUtil;
-import com.LetMeDoWith.LetMeDoWith.infrastructure.feed.query.FeedCacheRepository;
+import com.LetMeDoWith.LetMeDoWith.infrastructure.feed.cache.FeedCacheQueryRepository;
 import com.LetMeDoWith.LetMeDoWith.infrastructure.feed.query.FeedQueryRepository;
 import com.LetMeDoWith.LetMeDoWith.infrastructure.feed.query.dto.FeedbackAvailableDowithTaskQueryDto;
 import java.util.List;
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Service;
 public class FeedService {
 
     private final FeedQueryRepository feedQueryRepository;
-    private final FeedCacheRepository feedCacheRepository;
+    private final FeedCacheQueryRepository feedCacheQueryRepository;
 
     /**
      * 잔소리 대상 두윗 목록 조회. Redis에 적재된 두윗 목록을 조회하며, 실패시 현재시간을 기준으로 잔소리 대상인 두윗을 가져옴
@@ -22,21 +22,21 @@ public class FeedService {
      * @return 잔소리 대상 두윗
      */
     public RetrieveFeedbackAvailableDowithTasksResult retrieveFeedbackAvailableDowithTasks() {
-        List<Long> dowithTaskIds = feedCacheRepository.getFeedbackAvailableDowithTaskIds();
+        List<Long> dowithTaskIds = feedCacheQueryRepository.getFeedbackAvailableDowithTaskIds();
 
         // fallback
         if (dowithTaskIds.isEmpty()) {
             return RetrieveFeedbackAvailableDowithTasksResult.from(
-                    feedQueryRepository.getFeedbackAvailableDowithTasks(SystemTimeUtil.now()));
+                feedQueryRepository.getFeedbackAvailableDowithTasks(SystemTimeUtil.now()));
         }
 
         List<FeedbackAvailableDowithTaskQueryDto> feedbackAvailableDowithTasks =
-                feedCacheRepository.getFeedbackAvailableDowithTasks(dowithTaskIds);
+            feedCacheQueryRepository.getFeedbackAvailableDowithTasks(dowithTaskIds);
 
         // fallback
         if (feedbackAvailableDowithTasks.isEmpty()) {
             return RetrieveFeedbackAvailableDowithTasksResult.from(
-                    feedQueryRepository.getFeedbackAvailableDowithTasks(SystemTimeUtil.now()));
+                feedQueryRepository.getFeedbackAvailableDowithTasks(SystemTimeUtil.now()));
         }
 
         return RetrieveFeedbackAvailableDowithTasksResult.from(feedbackAvailableDowithTasks);
